fix: names fixed

+++ src/main/java/co/edu/unicauca/SIRENABackend/repositories/IReservaRepository.java
@@ -1,10 +1,11 @@
 package co.edu.unicauca.SIRENABackend.repositories;
 
-import co.edu.unicauca.SIRENABackend.models.ReservaModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import co.edu.unicauca.SIRENABackend.models.ReservaModel;
+
 @Repository
-public interface IReservaRepository extends JpaRepository<ReservaModel, Long> {
+public interface IReservaRepository extends JpaRepository<ReservaModel, Integer> {
     
 }
